2 п. для MyTask ДЗ 6 реализован в MyTaskService.

+++ src/main/java/ru/geekbrains/pattern/lessons/services/MyTaskService.java
@@ -4,18 +4,24 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.geekbrains.pattern.lessons.persistence.entities.MyTask;
-import ru.geekbrains.pattern.lessons.persistence.entities.utils.interfaces.ICreator;
 import ru.geekbrains.pattern.lessons.persistence.repositories.MyTaskRepository;
 
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
 public class MyTaskService {
     private final MyTaskRepository myTaskRepository;
+    private Map<Long, MyTask> identityMap = new HashMap<>();
 
-    public List<MyTask> showTasks() {
-        return myTaskRepository.findAll();
+    public Map<Long, MyTask> showTasks() {
+        if (identityMap.size() == 0){
+            for (MyTask task: myTaskRepository.findAll()) {
+                identityMap.put(task.getId(),task);
+            }
+        }
+        return identityMap;
     }
 
     public String addTask(){
@@ -30,12 +36,18 @@ public class MyTaskService {
                 .name("новая задача")
                 .build();
         myTaskRepository.save(task);
+        long id = myTaskRepository.findAll().indexOf(task);
+        identityMap.put(id,myTaskRepository.findById(id).get());
     }
 
     public String removeTask(Long id){
-        if (myTaskRepository.existsById(id)) {
+
+        if (identityMap.containsKey(id)){
+            identityMap.remove(id);
+            if (myTaskRepository.existsById(id)) {
             myTaskRepository.deleteById(id);
-        } // if not exist throw exception
+            }
+        }// if not exist throw exception
 
         return "redirect:/";
     }
